pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "buvis-pybase"
version = "0.5.7"
description = "Foundation of BUVIS python projects"
authors = [{ name = "Tomáš Bouška", email = "tomas@buvis.net" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "inflection>=0.5.1,<1",
  "jira>=3.10.5,<4",
  "pexpect>=4.9.0,<5",
  "pyyaml>=6.0.3,<7",
  "rake-nltk>=1.0.6,<2",
  "rich>=14.2.0,<15",
  "torch>=2.9.1,<3",
  "transformers>=4.57.3,<5",
  "tzlocal>=5.3.1,<6",
  "pywin32==311; sys_platform == 'win32'",
  "xattr>=1.3.0,<2; sys_platform != 'win32'",
]

[dependency-groups]
dev = ["pre-commit>=4.5.1,<5", "bump-my-version>=0.31,<1"]
test = ["pytest>=9.0.2,<10", "pytest-mock>=3.15.1,<4"]
docs = ["sphinx>=8.2.3,<9", "sphinx-rtd-theme>=3.0.2,<4"]

[tool.hatch.build.targets.wheel]
packages = ["src/buvis"]

[tool.uv]
add-bounds = "major"

[[tool.uv.index]]
name = "test-pypi"
url = "https://test.pypi.org/simple/"
explicit = true

[tool.bumpversion]
current_version = "0.5.7"
parse = """(?x)
    (?P<major>0|[1-9]\\d*)\\.
    (?P<minor>0|[1-9]\\d*)\\.
    (?P<patch>0|[1-9]\\d*)
    (?:-(?P<pre_l>dev|rc)(?P<pre_n>0|[1-9]\\d*))?
"""
serialize = [
  "{major}.{minor}.{patch}-{pre_l}{pre_n}",
  "{major}.{minor}.{patch}",
]
commit = true
tag = true
tag_name = "v{new_version}"
message = "chore(release): bump to {new_version}"
pre_commit_hooks = ["uv run pytest"]

[tool.bumpversion.parts.pre_l]
values = ["dev", "rc", "final"]
optional_value = "final"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101",    # assert is expected in tests
  "ANN",     # type annotations not required in tests
  "ARG",     # unused args are often fixtures
  "SLF001",  # private member access allowed in tests
  "PLR2004", # magic values ok in tests
]
